requirements.txt
streamlit>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pypdf>=3.15.1
matplotlib>=3.7.0
pandas>=1.5.3
numpy>=1.23.5
scikit-learn>=1.0.2
plotly>=5.14.0
faiss-cpu>=1.7.4
pillow>=9.4.0
requests>=2.28.2
datetime>=5.0
markdownify>=0.11.6